refactor: rename Global_HTMLElements to GlobalHTMLElements
- Renamed `Global_HTMLElements.ts` to `GlobalHTMLElements.ts`.
- Updated all imports to reflect the new file name.
