+++ src/Modules/GameStep2_inGame/GameBoard/GameBoard.ts
@@ -2,7 +2,7 @@ import {
   GlobalGameData,
   AutoRollPlayer
 } from '../../Global';
-import { ScoreBoardHolder } from '../../Global_HTMLElements';
+import { ScoreBoardHolder } from '../../GlobalHTMLElements';
 import { PlayerObj } from '../../../Classes/CreatePlayers';
 import { ScoreboardCard } from '../../../Templates/Templates';
 import { PlayersRoll } from '../PlayersRoll/PlayersRoll';
